Adding an option to an hook
As stated in
https://github.com/ankitects/anki/pull/438#issuecomment-583168521 I'm
trying to change anki (as little as possible) to remove monkey
patching from my add-ons.

Add-on 1983204951 does monkey patching because it needs to take into
account whether the export window is called from browser or
not. Ideally, I'd like that, when I call the export window, I can add
an option. This option will then be passed to exporters_list_created
to decide how to edit the list. (In the present case, it means:
whether or not to add "export selected card)

Since exporters_list_created calls the legacy hook exportersList, I
can't directly edit it. So I prefer to add a new hook.