DeckId
There is an inconsistency in the use of deck_id. Sometime as int,
sometime as string. Since the string always contains a int, using int
seems okay.

Furthermore, in schedv2, the comment
"# early alphas were setting the active ids as a str"
seems to indicate that the use of str was an accident

Anyway, I ensure conversion to type deck_id everywhere to stay
compatible with everything