hooks for class creation
It often occurs that add-on must redefine the `__init__` method in order to do one last action when class is loaded. This PR create a hook `class_name_did_init` which is called when a method init is called on the class class_name.

I decided to create a simple class which redefine `__init__`, in order to avoid code repetition. Actually, it allows to create a new line in each `__init__` AND adding one hook by class in pylib/tools/genhooks.py. The biggest downside is that any add-on dev' which looks for this hook won't directly see it in the class. So this need to be properly documented in the add-on creation manual

Creating a real meta class for HookOnInit would have been nice too. The unexpected problem here wath that each PyQt class have already their meta class. Which led to a meta class conflict when I added HookOnInit to any class which was a child of a QTClass. This is the reason why I simply used __init_subclass__, which lead to pylint saying W0231. I know that meta class can be merged, but I didn't want to venture in such complex python programming.

For example
```python
from anki import hooks
def init_card(card):
    print(card.id)
hooks.Card_did_init.append(init_card)
```
would print the id of each card created.