Reset can be cancelled
I believe that reset is called far too often. When a card is not
modified, it can still be displayed. It's actually sometime useful
when the content of the current card made me know I need to change
another card.
I want to be able to cancel the reset. Even if it means that the
scheduler will not be rebuild and thus in some case this change which
card I see