add switch in prefs for new timezone handling code; bump sync version
AD devs: before updating the sync version to 10, please make sure
the new timezone code has been ported over. The core change to the
scheduler is in _updateCutoff():

https://github.com/ankitects/anki/blob/b198ba8073adf3b18e30ae4fe6c5c81a7e52e2e0/pylib/anki/schedv2.py#L1357

with the following supporting Python functions:

https://github.com/ankitects/anki/blob/b198ba8073adf3b18e30ae4fe6c5c81a7e52e2e0/pylib/anki/schedv2.py#L1410

 _timing_today() is calculated in Rust (lines up to 92, the rest are
tests):

https://github.com/ankitects/anki/blob/b198ba8073adf3b18e30ae4fe6c5c81a7e52e2e0/rslib/src/sched/cutoff.rs#L20

The change went through a few iterations before stabilising, so it's
probably easier to refer to the above code than the patches that got
us to that point.