normalize name in _checkDeckTree
Currently, if there is two decks "A" and "a", they won't be considered
as duplicate. So they won't be corrected.
That's hard to test, because you're not supposed to be able to create
those two decks. However, it seems that it may still be possible on
AnkiDroid (correction is in master, and probabla alpha)

Thus, I suggest to compare normalized names instead of names
themselves.
Actually, I'm even surprised that I didn't do this PR while doing dae/anki@e44b0492781e8248318e87a47eb179561db94b7b