hook importer_does_it_import_note
The second argument should be a tuple. But since note is a list, I
choosed to keep it as a list, even if it means I can't give its
precise type.

The main reason for this add-on is so that it can considerably
simplify the add-on Special Fields, which is currently quite a
horrible monkey patch