hook importer_change_update
Once again, this would be useful in "special fields" add-ons. I do
realize it seems redundant with importer_does_it_update_note, but
actually I don't consider it to be the case. I expect
importer_change_update to change the notes themselves, and not
necessarily to add/remove elements to the list.