Hook empty_card_line
I expect to use this with my add-on 25425599. This add-on explain
which are the name of the cards that are going to be deleted, instead
of only their ord. I find it quite clearer when I just changed a card
type and want to make sure that the correct cards will be removed/try
to understand a bug.

I believe this add-on should be incorpated in anki by the way, it's
simple and quite more helpful