Editor: on_focus_hook
Add an hook that will be triggered each time focus is put in a
field. I expect to use this in my addon 882784122. In this add-on, I
display mathjax/latex result in editor unless the focus is in the
editor. It will be simpler for me to just have this look and tell my
add-on to add an element to the hook than to add a new "focus" element
to each tag each time.

I expect it to be particularly useful because it will ensure that the
hook is filled after editor.ts is loaded but before the fuction
`loadNote` is called, so that the hook is even called on the very
first "focusField"