Editor: on_send_field_content_hook
This hook allow to change the field content that is sent to Python.
In my add-on "resize image" it is simpler for me to have the change
done in JS than in python. I thus would prefer to have a hook here.

Furthermore, I want to change the value sent, not the field
itself. Because I want to remove the html part which deals with
resizing images and not put it in the field content. However, I want
the user to still be able to resize if they want.