Editor: Hook name_line_hook
My ad-ons:
* clean html
* compatible frozen field
* multi column window

All adds buttons near to the fields name. This ensure that it is easy
to apply an action to a specific field without needing to first put
the focus on the field and then find the correct shortuct/menu item.

Current, I am looking for the td #name${i}. However, this hook would
allow to simplify code by ensuring that I can create the buttons more
easily.