Remove code duplication _toBury
The methods _toBury have a lot of code duplicated.

The reason behind this PR is that AnkiDroid tries to stay close to
Anki. However, for efficiency purpose, I need to change the burying
part. This mean that I'll add even more code duplication. Having a
function would thus starts to save a lot of space and avoid errors.