Fix late blur causing IndexError
Changing note type to another type with less fields using Ctrl+N throws an exception
if a field in the original type with an ordinal higher than the maximum field ordinal
of the new type was focused prior to the note type change.
This is due to a late blur processing in onBridgeCmd.