Revert "Merge pull request #873 from hgiesel/otherjsdeps"
This reverts commit ad7491a9ae49101265b52a7d8a5a109f68142cd1, reversing
changes made to 5a79515f2ae46129e41fa217b41b8c6e54efb94b.

- Standard graphs render incorrectly on latest version - the wrong number
of days are shown, and the grid lines look wrong. Any version after 0.8.3
seems to suffer from this problem.
- Pie graphs and stack graphs don't render - they are provided in separate
files, and plot.js in previous Anki versions has them included in the one
file. To maintain compatibility with add-ons, we'd need to create a single
file as before, instead of importing multiple files.

If the above issues are fixed I'd be happy to merge this in again, but
as the old graphs are on the way out, it's probably not worth the effort.