Revert "Merge pull request #873 from hgiesel/otherjsdeps"
This reverts commit 62600051ae1415bf9c41503a5f6079519e92af4b, reversing
changes made to 88553acb0d37fa6a46f7a917acec4d33a321427d.

- Standard graphs render incorrectly on latest version - the wrong number
of days are shown, and the grid lines look wrong. Any version after 0.8.3
seems to suffer from this problem.
- Pie graphs and stack graphs don't render - they are provided in separate
files, and plot.js in previous Anki versions has them included in the one
file. To maintain compatibility with add-ons, we'd need to create a single
file as before, instead of importing multiple files.

If the above issues are fixed I'd be happy to merge this in again, but
as the old graphs are on the way out, it's probably not worth the effort.